Trim macOS app architecture indirection
